fixed problem n+1 in /patiens

+++ src/main/java/com/doctorvisits/visittracker/dto/DoctorDataDto.java
@@ -15,5 +15,5 @@ import lombok.experimental.FieldDefaults;
 public class DoctorDataDto {
     String firstName;
     String lastName;
-    int totalPatients;
+    Long totalPatients;
 }
